requirements.txt
striprtf
python-docx
